requirements.txt
# Traffic Light Detection System - Requirements
# Python 3.8+

# Deep Learning & Computer Vision
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Web Application Framework
streamlit>=1.28.0

# YOLOv5 Dependencies
ultralytics
PyYAML>=5.3.1
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Utilities
requests>=2.31.0
